rename classes and create shared class for fast-forward merge action
