+++ frontend/actions/src/main/java/com/virtuslab/gitmachete/frontend/actions/contextmenu/PullSelectedBranchAction.java
@@ -3,10 +3,10 @@ package com.virtuslab.gitmachete.frontend.actions.contextmenu;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import io.vavr.control.Option;
 
-import com.virtuslab.gitmachete.frontend.actions.base.BasePullBranchFastForwardOnlyAction;
+import com.virtuslab.gitmachete.frontend.actions.base.BasePullBranchAction;
 import com.virtuslab.gitmachete.frontend.actions.expectedkeys.IExpectsKeySelectedBranchName;
 
-public class PullSelectedBranchFastForwardOnlyAction extends BasePullBranchFastForwardOnlyAction
+public class PullSelectedBranchAction extends BasePullBranchAction
     implements
       IExpectsKeySelectedBranchName {
   @Override
