+++ frontend/base/src/main/resources/GitMacheteBundle.properties
@@ -66,11 +66,11 @@ action.GitMachete.BaseOverrideForkPointAction.dialog.override-fork-point.radio-b
 action.GitMachete.BaseOverrideForkPointAction.dialog.override-fork-point.radio-button.inferred=Inferred commit ({0})
 
 
-action.GitMachete.BasePullBranchFastForwardOnlyAction.action-name=P_ull
-action.GitMachete.BasePullBranchFastForwardOnlyAction.description-action-name=Pull (fast-forward only)
-action.GitMachete.BasePullBranchFastForwardOnlyAction.notification.title.pull-fail=Pull of branch ''{0}'' failed
-action.GitMachete.BasePullBranchFastForwardOnlyAction.notification.title.pull-success=Branch <b>{0}</b> has been pulled
-action.GitMachete.BasePullBranchFastForwardOnlyAction.task-title=Fetching
+action.GitMachete.BasePullBranchAction.action-name=P_ull
+action.GitMachete.BasePullBranchAction.description-action-name=Pull (fast-forward only)
+action.GitMachete.BasePullBranchAction.notification.title.pull-fail=Pull of branch ''{0}'' failed
+action.GitMachete.BasePullBranchAction.notification.title.pull-success=Branch <b>{0}</b> has been pulled
+action.GitMachete.BasePullBranchAction.task-title=Fetching
 
 
 action.GitMachete.BasePushBranchAction.action-name=_Push
@@ -201,9 +201,9 @@ action.GitMachete.OverrideForkPointOfSelectedBranchAction.GitMacheteContextMenu.
 action.GitMachete.OverrideForkPointOfSelectedBranchAction.description=Select after which commit the range of commits belonging to this branch starts
 
 
-action.GitMachete.PullSelectedBranchFastForwardOnlyAction.text=Git Machete: Pull Selected Branch
-action.GitMachete.PullSelectedBranchFastForwardOnlyAction.GitMacheteContextMenu.text=P_ull
-action.GitMachete.PullSelectedBranchFastForwardOnlyAction.description=Pull (fast-forward only) this branch
+action.GitMachete.PullSelectedBranchAction.text=Git Machete: Pull Selected Branch
+action.GitMachete.PullSelectedBranchAction.GitMacheteContextMenu.text=P_ull
+action.GitMachete.PullSelectedBranchAction.description=Pull (fast-forward only) this branch
 
 
 action.GitMachete.PushSelectedBranchAction.text=Git Machete: Push Selected Branch...
@@ -352,10 +352,9 @@ action.GitMachete.OverrideForkPointOfCurrentBranchAction.text=Git Machete: Overr
 action.GitMachete.OverrideForkPointOfCurrentBranchAction.GitMacheteToolbar.text=Override Fork Point of Current Branch...
 action.GitMachete.OverrideForkPointOfCurrentBranchAction.description=Select after which commit the range of commits belonging to the current branch starts
 
-
-action.GitMachete.PullCurrentBranchFastForwardOnlyAction.text=Git Machete: Pull Current Branch
-action.GitMachete.PullCurrentBranchFastForwardOnlyAction.GitMacheteToolbar.text=Pull Current Branch
-action.GitMachete.PullCurrentBranchFastForwardOnlyAction.description=Pull (fast-forward only) current branch
+action.GitMachete.PullCurrentBranchAction.text=Git Machete: Pull Current Branch
+action.GitMachete.PullCurrentBranchAction.GitMacheteToolbar.text=Pull Current Branch
+action.GitMachete.PullCurrentBranchAction.description=Pull (fast-forward only) current branch
 
 
 action.GitMachete.PushCurrentBranchAction.text=Git Machete: Push Current Branch...
